+++ CHANGELOG.md
@@ -3,30 +3,37 @@
 ## [Unrealeased] - yyyy-mm-dd
 
 ### Added
- 
+
+- Update task status in dp when copy
+
 ### Changed
- 
+
 ### Fixed
+
 - Fix Pipeline type not found bug
 - Remove unused gitlab method
 
 ## [1.6.0] - 2020-12-31
 
 ### Changed
+
 - Show pipeline failed message in tracker
 
 ## [1.5.0] - 2020-12-31
 
 ### Changed
+
 - Move todoConfigChoices to config file
 - Use environment variable to set config file path
 
 ## [1.4.0] - 2020-12-30
 
 ### Added
+
 - Add output message for end command
- 
+
 ### Changed
+
 - Do not open issue page when start
 - Change notification message of tracker
 - Add "debug" choice to todoConfig
@@ -34,20 +41,24 @@
 ## [1.3.0] - 2020-12-30
 
 ### Fixed
+
 - Use commit hash and branch ref to search pipeline to avoid cannot get correct pipeline bug
 
 ## [1.2.0] - 2020-12-28
 
 ### Added
+
 - Use template to manage todo list
 
 ## [1.1.0] - 2020-12-28
 
 ### Added
+
 - Remove item when track task is finished
 - Use deploy job status to check deploy stage is finished
 
 ## [1.0.0] - 2020-12-23
 
 ### Added
+
 - Init project